mempool_malloc and free now handles pointers starts from mempool_header.

+++ src/arch/util/mempool.h
@@ -30,7 +30,7 @@ typedef struct mempool_header
   size_t            next_free;
 } mempool_header;
 
-typedef void * (* mempool_newblockfn)(int *size, gni_mem_handle_t *mem_hndl);
+typedef void * (* mempool_newblockfn)(size_t *size, gni_mem_handle_t *mem_hndl);
 typedef void (* mempool_freeblock)(void *ptr, gni_mem_handle_t mem_hndl);
 
 // only at beginning of first block of mempool
